Refactor Visitor into its own package
Fix #25
