--- fluent.syntax/src/test/java/org/projectfluent/syntax/ast/JavaVisitorTest.java
@@ -1,41 +1,0 @@
-package org.projectfluent.syntax.ast;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import org.junit.jupiter.api.Test;
-import org.projectfluent.syntax.parser.FluentParser;
-
-import java.util.regex.Pattern;
-
-class JavaWordCounter extends Visitor {
-    private static final java.util.regex.Pattern WORD_BOUNDARY = Pattern.compile("\\W+");
-
-    private int words;
-    public JavaWordCounter() {
-        words = 0;
-    }
-
-    public int getWords() {
-        return words;
-    }
-
-    void visitResource(Resource node) {
-        System.out.println("resource");
-        this.genericVisit(node);
-    }
-
-    public void visitTextElement(TextElement node) {
-        String val = node.getValue();
-        words += WORD_BOUNDARY.split(val).length;
-    }
-}
-
-class JavaVisitorTest {
-    private final FluentParser parser = new FluentParser(false);
-
-    @Test
-    void test_wordcounter() {
-        Resource res = parser.parse("msg = value with words");
-        JavaWordCounter counter = new JavaWordCounter();
-        counter.visit(res);
-        assertEquals(3, counter.getWords());
-    }
-}
